ml_service/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.1.0

# Database Connectivity
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3

# Machine Learning Libraries
scikit-learn==1.5.2
numpy==1.26.4
scipy==1.15.3
joblib==1.5.1

# Text Processing and NLP
nltk==3.9.1

# Data Visualization and Analysis
matplotlib==3.9.3
seaborn==0.13.2
plotly

# HTTP Requests
httpx==0.28.1
requests==2.32.3

# Date and Time Handling
python-dateutil==2.9.0.post0

# Logging and Utilities
rich==14.0.0
click==8.2.1

# Development and Debugging
pydantic==2.10.3

# Enhanced ML and Visualization
umap-learn==0.5.6
shap==0.46.0
lime==0.2.0.1

# Transformer Models for Semantic Understanding
transformers==4.36.2
torch==2.7.0
sentence-transformers==2.2.2

# System Monitoring and Testing
psutil==6.1.1
aiohttp==3.9.1

# Testing and Evaluation
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
bandit==1.7.5
safety>=3.0.0

# Additional ML Libraries for Advanced Features
imbalanced-learn==0.11.0
xgboost==2.0.2
lightgbm==4.1.0

# Additional dependencies for compatibility
faker==28.1.0 